+++ love/src/components/ScriptQueue/Scripts/ContextMenu/ContextMenu.jsx
@@ -3,7 +3,9 @@ This file is part of LOVE-frontend.
 
 Copyright (c) 2023 Inria Chile.
 
-Developed by Inria Chile.
+Developed by Inria Chile and the Telescope and Site Software team.
+
+Developed for the Vera C. Rubin Observatory Telescope and Site Systems.
 
 This program is free software: you can redistribute it and/or modify it under 
 the terms of the GNU General Public License as published by the Free Software 
@@ -19,18 +21,10 @@ this program. If not, see <http://www.gnu.org/licenses/>.
 
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-
 import styles from './ContextMenu.module.css';
 
 export default class ContextMenu extends PureComponent {
   static propTypes = {
-    /**Position data for the context menu.
-     * Usually from event.target.getBoundingClientRect()
-     */
-    contextMenuData: PropTypes.shape({
-      right: PropTypes.number,
-      bottom: PropTypes.number,
-    }),
     isOpen: PropTypes.bool,
     /** List of clickable options to be displayed */
     options: PropTypes.arrayOf(
@@ -50,7 +44,6 @@ export default class ContextMenu extends PureComponent {
   };
 
   static defaultProps = {
-    contextMenuData: {},
     isOpen: false,
     options: [],
     target: undefined,
@@ -63,9 +56,9 @@ export default class ContextMenu extends PureComponent {
     };
   }
 
-  componentDidUpdate = (nextState, nextProps) => {
-    if (this.props.target !== nextProps.target) {
-      const parentCustomView = this.props.target ? this.props.target.closest('.react-grid-item') : undefined;
+  componentDidUpdate = (prevProps) => {
+    if (prevProps.target !== this.props.target) {
+      const parentCustomView = this.props.target?.closest('.react-grid-item');
       const offset = parentCustomView ? parentCustomView.getBoundingClientRect().x : 0;
       this.setState({
         offset,
@@ -74,29 +67,37 @@ export default class ContextMenu extends PureComponent {
   };
 
   render() {
+    const { isOpen, options, target, children } = this.props;
+    const { offset } = this.state;
+    const targetBoundingRect = target ? target.getBoundingClientRect() : { right: 0, bottom: 0 };
+
     return (
-      this.props.isOpen && (
+      isOpen && (
         <div
           className={styles.container}
           style={{
-            left: this.props.contextMenuData.right - this.state.offset,
-            top: `calc( -3.3em + ${this.props.contextMenuData.bottom}px)`,
+            left: targetBoundingRect.right - offset,
+            top: `calc( -3.3em + ${targetBoundingRect.bottom}px)`,
+            width: children ? '20em' : 'auto',
           }}
           onClick={(e) => e.stopPropagation()}
           onMouseDown={(e) => e.stopPropagation()}
         >
-          {this.props.options.map((child, index) => {
-            return (
-              <div
-                className={[styles.row, child.disabled ? '' : styles.enabled].join(' ')}
-                key={index}
-                onClick={!child.disabled ? child.action : undefined}
-              >
-                <div className={[styles.iconWrapper].join(' ')}>{child.icon}</div>
-                <div className={[styles.buttonText].join(' ')}>{child.text}</div>
-              </div>
-            );
-          })}
+          <div className={styles.options}>
+            {options.map((child, index) => {
+              return (
+                <div
+                  className={[styles.row, child.disabled ? '' : styles.enabled].join(' ')}
+                  key={index}
+                  onClick={!child.disabled ? child.action : undefined}
+                >
+                  <div className={[styles.iconWrapper].join(' ')}>{child.icon}</div>
+                  <div className={[styles.buttonText].join(' ')}>{child.text}</div>
+                </div>
+              );
+            })}
+          </div>
+          {children}
         </div>
       )
     );