+++ CHANGELOG.rst
@@ -2,6 +2,19 @@
 Version History
 ===============
 
+v6.11.1
+-------
+
+* Update package.json version to v6.11.1. `<https://github.com/lsst-ts/LOVE-frontend/pull/775>`_
+* Fix CSCDetail & CSCExpanded containers to subscribe to offlineDetailedState event only for big cameras: ATCamera, MTCamera and CCCamera. `<https://github.com/lsst-ts/LOVE-frontend/pull/773>`_
+* Improvements for the observatory status feature. `<https://github.com/lsst-ts/LOVE-frontend/pull/772>`_
+
+v6.11.0
+-------
+
+* Update package.json version to v6.11.0. `<https://github.com/lsst-ts/LOVE-frontend/pull/771>`_
+* Implement observatory status feature. `<https://github.com/lsst-ts/LOVE-frontend/pull/770>`_
+
 v6.10.2
 -------